tools/mp3-config-generator/pyproject.toml
[project]
name = "mp3-config-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.7",
    "pyyaml>=6.0.1",
]
